Add executable prototype.
